+++ src/helix/request.rs
@@ -436,3 +436,35 @@ pub trait RequestGet: Request {
         ))
     }
 }
+
+/// Parses a response where Helix responds with a single datum inside `data`.
+///
+/// An example response is `{ "data": [ { "foo": 1 } ]`.
+pub(crate) fn parse_single_return<T, E>(
+    request: Option<T>,
+    uri: &http::Uri,
+    response: &str,
+    status: http::StatusCode,
+) -> Result<Response<T, T::Response>, E>
+where
+    T: Request,
+    E: errors::HelixRequestDeserError,
+{
+    let resp = match status {
+        http::StatusCode::OK => {
+            let resp: InnerResponse<[T::Response; 1]> = parse_json(response, true)
+                .map_err(|e| E::deserialize_error(response.to_string(), e, uri.clone(), status))?;
+            let [s] = resp.data;
+            s
+        }
+        _ => {
+            return Err(E::invalid_response(
+                "unexpected status code",
+                response.to_string(),
+                status,
+                uri.clone(),
+            ));
+        }
+    };
+    Ok(Response::with_data(resp, request))
+}