+++ CHANGELOG.md
@@ -79,6 +79,7 @@
 - Added beta `channel.ad_break.begin` eventsub event
 - Added `conduit.shard.disable` EventSub event
 - Added `title` and `description` as fields in the response of `Get Channel Chat Badges` and `Get Global Chat Badges`
+- Added `Get AutoMod Settings` and `Update AutoMod Settings` endpoints
 
 ### Fixed